+++ CHANGELOG.md
@@ -1,7 +1,18 @@
+## 0.0.20
+
+* [#f8d7d33](https://github.com/dart-lang/sdk/commit/f8d7d3300c44202001126de33f0145e7055d1ed8):
+  Do not close _HttpClientConnection when the HTTP parse stream ends, if it has already been closed.
+* [#e8c0353](https://github.com/dart-lang/sdk/commit/f753fe4144ad0939f15c577edc08f274939928de):
+  Throw RedirectException when a redirect is missing a Location header
+* [#e8c0353](https://github.com/dart-lang/sdk/commit/e8c035308bec001af477d644f3a5c396dedc94ff):
+  [VM/Runtime] Fix writeln to ensure the entire write is atomic
+* [#6edc0c5](https://github.com/dart-lang/sdk/commit/6edc0c580db1c8aa0034cf736585b7a1c1bc1be2):
+  _http: correct some unresolvable comment references
+
 ## 0.0.19
 
-* [#35a595f](https://github.com/dart-lang/sdk/commit/35a595f66b2f55deb88e88f12fb7b89e622d26cc):
-  [io] [beta] [ dart:io ] Fix minor bugs in network profiler
+* [#791163e](https://github.com/dart-lang/sdk/commit/791163eef8dd8d2deaaaf5e9c3ecb96308d8a7b3):
+  [ dart:io ] Fix minor bugs in network profiler
 * [#fc4bcab](https://github.com/dart-lang/sdk/commit/fc4bcab2efd43c8928c26d66a194785f5fd52ce4):
   [io] Propagate cancellation in _HttpOutgoing.addStream